+++ lib/browser.js
@@ -2,16 +2,17 @@
 "use strict";
 
 const path = require('path');
-const ErrorFactory = require('../errors/error_factory');
+const ErrorFactory = require('./errors/error_factory');
 const BrowserBase = require('./browser_base');
-const BrowserAssertions = require('../assertions/browser_assertions');
-const LocalStorageAssertions = require('../assertions/local_storage_assertions');
-const BrowserLocalStorage = require('./browser_local_storage');
-const BrowserCookies = require('./browser_cookies');
-const utils = require('../utils');
+const BrowserAssertions = require('./modules/assertions/browser_assertions');
+const LocalStorageAssertions = require('./modules/assertions/local_storage_assertions');
+const BrowserLocalStorage = require('./modules/browser_local_storage');
+const BrowserCookies = require('./modules/browser_cookies');
+const utils = require('./utils');
+const config = require('../config');
 
-const injectionScriptsPath = path.join(__dirname, "../..", "injection_scripts");
-const injectionScripts = ["selector_query.js", "wendigo_utils.js"];
+const injectionScriptsPath = config.injectionScripts.path;
+const injectionScripts = config.injectionScripts.files;
 
 
 module.exports = class Browser extends BrowserBase {
