+++ lib/browser_base.js
@@ -1,7 +1,7 @@
 /* global WendigoUtils */
 "use strict";
 
-const ErrorFactory = require('../errors/error_factory');
+const ErrorFactory = require('./errors/error_factory');
 
 function pageLog(log) {
     console[log._type](log._text); // eslint-disable-line
