+++ lib/modules/assertions/browser_assertions.js
@@ -5,12 +5,13 @@ const BrowserNotAssertions = require('./browser_not_assertions');
 const elementsAssertionUtils = require('./utils/assert_elements');
 const textUtils = require('./utils/text_utils');
 const assertUtils = require('./utils/assert_utils');
-const ErrorFactory = require('../errors/error_factory');
+const ErrorFactory = require('../../errors/error_factory');
+const BrowserModule = require('../browser_module');
 
-module.exports = class BrowserAssertions {
+module.exports = class BrowserAssertions extends BrowserModule {
 
     constructor(browser) {
-        this._browser = browser;
+        super(browser);
         this.not = new BrowserNotAssertions(this, browser);
     }
