+++ lib/modules/assertions/local_storage_assertions.js
