+++ lib/modules/assertions/utils/assert_utils.js
@@ -1,7 +1,7 @@
 "use strict";
 
-const Errors = require('../../errors/errors');
-const ErrorFactory = require('../../errors/error_factory');
+const Errors = require('../../../errors/errors');
+const ErrorFactory = require('../../../errors/error_factory');
 
 module.exports = {
     invertify(cb, msg) {
