+++ lib/modules/assertions/utils/text_utils.js
