+++ lib/modules/browser_module.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = class BrowserModule {
+    constructor(browser) {
+        this._browser = browser;
+    }
+};
